# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a stock management system.",
        "title": "GoStock API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Retorna uma lista de produtos com base em filtros e paginação.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Lista produtos com filtros e paginação",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número da página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limite de itens por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por nome do produto",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por SKU",
                        "name": "sku",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar apenas por produtos ativos",
                        "name": "active_only",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Parâmetros de query inválidos",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria um novo produto principal e suas variantes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Cria um novo produto com variantes",
                "parameters": [
                    {
                        "description": "Dados do produto e variantes",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "400": {
                        "description": "Payload inválido",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Busca um produto específico e suas variantes pelo ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Obtém um produto por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stock/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza a quantidade de estoque para uma variante de produto em um armazém específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Ajusta o nível de estoque de um produto em um armazém",
                "parameters": [
                    {
                        "description": "Dados para ajuste de estoque",
                        "name": "adjustment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.StockAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nível de estoque atualizado",
                        "schema": {
                            "$ref": "#/definitions/domain.StockLevel"
                        }
                    },
                    "400": {
                        "description": "Payload inválido",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflito de concorrência (versão)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.ErrorResponse": {
            "description": "Estrutura padronizada para respostas de erro na API.",
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "O nome do armazém não pode ser vazio."
                }
            }
        },
        "domain.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "sku": {
                    "description": "Stock Keeping Unit (código único de produto)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variants": {
                    "description": "Relações (serão gerenciadas por outras entidades/serviços)\nVariantes []Variant // Exemplo: Lista de tamanhos/cores",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Variant"
                    }
                }
            }
        },
        "domain.StockAdjustmentRequest": {
            "type": "object",
            "required": [
                "delta",
                "variant_id",
                "warehouse_id"
            ],
            "properties": {
                "delta": {
                    "description": "Quantidade a ser adicionada/removida",
                    "type": "integer"
                },
                "variant_id": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
        "domain.StockLevel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "variant_id": {
                    "type": "string"
                },
                "version": {
                    "description": "Para Controle de Concorrência Otimista (OCC)",
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.Variant": {
            "type": "object",
            "properties": {
                "attribute": {
                    "description": "Ex: \"Cor\"",
                    "type": "string"
                },
                "barcode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price_diff": {
                    "description": "Ajuste de preço para esta variante",
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "value": {
                    "description": "Ex: \"Vermelho\"",
                    "type": "string"
                }
            }
        },
...
        "product.ProductCreateRequest": {
            "type": "object",
            "properties": {
                "Product": {
                    "$ref": "#/definitions/domain.Product"
                },
                "Variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Variant"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}